Consolidate conditional localhost resolution into existing file
