+++ BUILD.gn
@@ -324,8 +324,6 @@ config("grpc_config") {
         "src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_wrapper_libuv.cc",
         "src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_wrapper_posix.cc",
         "src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_wrapper_windows.cc",
-        "src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_wrapper_windows_inner.cc",
-        "src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_wrapper_windows_inner.h",
         "src/core/ext/filters/client_channel/resolver/dns/dns_resolver_selection.cc",
         "src/core/ext/filters/client_channel/resolver/dns/dns_resolver_selection.h",
         "src/core/ext/filters/client_channel/resolver/dns/native/dns_resolver.cc",
