+++ src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_wrapper.h
@@ -87,14 +87,6 @@ void grpc_ares_complete_request_locked(grpc_ares_request* request);
 /* E.g., return false if ipv6 is known to not be available. */
 bool grpc_ares_query_ipv6();
 
-/* Maybe (depending on the current platform) checks if "name" matches
- * "localhost" and if so fills in addrs with the correct sockaddr structures.
- * Returns a bool indicating whether or not such an action was performed.
- * See https://github.com/grpc/grpc/issues/15158. */
-bool grpc_ares_maybe_resolve_localhost_manually_locked(
-    const char* name, const char* default_port,
-    grpc_core::UniquePtr<grpc_core::ServerAddressList>* addrs);
-
 /* Sorts destinations in lb_addrs according to RFC 6724. */
 void grpc_cares_wrapper_address_sorting_sort(
     grpc_core::ServerAddressList* addresses);
